# Use cached study direction and trial for `_CachedStorage`'s `get_best_trial`