## Motivation
Currently,  `_CachedStorage`'s `get_best_trial` just calls the method of the backend storage (`RDBStorage`). However, performance can be improved by using cached study directions and trials.

## Description of the changes
- Introduce `_get_best_trial_id` in `RDBStorage`
- Use cached study direction and trial for `_CachedStorage`'s `get_best_trial`

## Benchmark

```python
import optuna

def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -100, 100)
    y = trial.suggest_int("y", -100, 100)
    return x**2 + y**2

sampler = optuna.samplers.TPESampler(seed=42, multivariate=True, constant_liar=True)
study = optuna.create_study(sampler=sampler, storage="sqlite:///tmp.db")
study.optimize(objective, n_trials=1000)

```

| master | PR |
| - | - |
| 10.02s | 9.00s |